+++ src/components/ListView/BsListNavItem.ts
@@ -9,21 +9,36 @@ import {
     onMounted,
     ref,
     shallowRef,
-    watchEffect
+    watchEffect,
 } from 'vue';
 import { useCurrentRoute, useHasRouter, useIsRouteMatch } from '../../mixins/CommonApi';
-import type { IListItem, IListViewProvider, TBsListNavItem, TListNavItemOptionProps, TRecord } from '../../types';
+import type {
+    IListItem,
+    IListViewProvider,
+    TBsListNavItem,
+    TListNavItemOptionProps,
+    TRecord,
+} from '../../types';
 import ListItem from './mixins/ListItem';
 import {
     useAddChild,
     useListNavItemClasses,
     useListNavItemInnerClasses,
     useNavItemContentStyles,
-    useRenderListNavItem
+    useRenderListNavItem,
 } from './mixins/listNavApi';
 import { listNavItemProps } from './mixins/listViewProps';
 
-export default defineComponent<TBsListNavItem, TRecord, TRecord, ComputedOptions, MethodOptions, ComponentOptionsMixin, ComponentOptionsMixin, EmitsOptions>({
+export default defineComponent<
+    TBsListNavItem,
+    TRecord,
+    TRecord,
+    ComputedOptions,
+    MethodOptions,
+    ComponentOptionsMixin,
+    ComponentOptionsMixin,
+    EmitsOptions
+>({
     name: 'BsListNavItem',
     props: listNavItemProps,
     emits: [
@@ -31,33 +46,32 @@ export default defineComponent<TBsListNavItem, TRecord, TRecord, ComputedOptions
         /**
          * Fired when this component's state is updated.
          */
-        'update:active'
+        'update:active',
     ],
-    setup(props, {emit, expose, slots}) {
+    setup(props, { emit, expose, slots }) {
         const cmpProps = props as Readonly<TListNavItemOptionProps>;
+        const instance = shallowRef(getCurrentInstance());
         const refItem = shallowRef<IListItem>();
         const isActive = ref<boolean | undefined>(cmpProps.active);
         const expanded = ref<boolean>(false);
         const hasChild = ref<boolean>(false);
         const hasRouter = ref<boolean>(false);
 
-        expose({isActive, expanded});
+        expose({ isActive, expanded });
 
         const provider = inject<IListViewProvider>('ListView');
-        const navItemClasses = computed<TRecord>(
-            () => useListNavItemClasses(cmpProps, isActive, expanded, hasChild)
+        const navItemClasses = computed<TRecord>(() =>
+            useListNavItemClasses(cmpProps, isActive, expanded, hasChild)
         );
-        const navItemInnerClasses = computed<TRecord>(
-            () => useListNavItemInnerClasses(cmpProps, isActive, hasRouter, provider)
+        const navItemInnerClasses = computed<TRecord>(() =>
+            useListNavItemInnerClasses(cmpProps, isActive, hasRouter, provider)
         );
-        const navItemInnerStyles = computed<string[]>(
-            () => useNavItemContentStyles(cmpProps)
-        );
+        const navItemInnerStyles = computed<string[]>(() => useNavItemContentStyles(cmpProps));
 
         if (useHasRouter(cmpProps)) {
             const route = useCurrentRoute();
             watchEffect(() => {
-                if (provider && route && useIsRouteMatch(route, cmpProps)) {
+                if (provider && route && useIsRouteMatch(instance, route, cmpProps)) {
                     provider.activeItem = refItem.value;
                     let parent = refItem.value?.parent;
                     while (parent) {
@@ -67,45 +81,57 @@ export default defineComponent<TBsListNavItem, TRecord, TRecord, ComputedOptions
                 }
             });
         }
-        onBeforeMount(
-            () => {
-                const instance = getCurrentInstance();
-                if (instance) {
-                    refItem.value = new ListItem(<string>props.id, 'BsListNavItem', instance, emit);
 
-                    if (provider) {
-                        nextTick().then(() => useAddChild(provider, instance?.parent, refItem.value));
-                    }
+        onBeforeMount(() => {
+            instance.value = getCurrentInstance();
+            if (instance) {
+                refItem.value = new ListItem(
+                    props.id as string,
+                    'BsListNavItem',
+                    instance.value!,
+                    emit
+                );
+
+                if (provider) {
+                    nextTick().then(() =>
+                        useAddChild(provider, instance.value?.parent, refItem.value)
+                    );
                 }
             }
-        );
-        onMounted(
-            () => {
-                hasRouter.value = useHasRouter(cmpProps);
-                if (hasRouter.value) {
-                    const route = useCurrentRoute();
-                    if (route && useIsRouteMatch(route, cmpProps)) {
-                        refItem.value?.setActive(true);
+        });
+        onMounted(() => {
+            hasRouter.value = useHasRouter(cmpProps);
+            if (hasRouter.value) {
+                const route = useCurrentRoute();
+                if (route && useIsRouteMatch(instance, route, cmpProps)) {
+                    refItem.value?.setActive(true);
+                }
+            }
+            nextTick().then(() => {
+                hasChild.value = refItem.value?.hasChild() ?? false;
+                if (hasRouter.value && !hasChild.value && isActive.value) {
+                    let parent = refItem.value?.parent;
+                    while (parent) {
+                        parent.setActive(true);
+                        provider?.expand(parent);
+                        parent = parent.parent;
                     }
                 }
-                nextTick().then(() => {
-                    hasChild.value = refItem.value != undefined && refItem.value.hasChild();
-                    if (hasRouter.value && !hasChild.value && isActive.value) {
-                        let parent = refItem.value?.parent;
-                        while (parent) {
-                            parent.setActive(true);
-                            provider?.expand(parent);
-                            parent = parent.parent;
-                        }
-                    }
-                })
-            }
-        )
+            });
+        });
 
         return () =>
             useRenderListNavItem(
-                slots, props, navItemClasses, navItemInnerClasses, navItemInnerStyles,
-                isActive, expanded, hasChild, refItem, provider,
+                slots,
+                props,
+                navItemClasses,
+                navItemInnerClasses,
+                navItemInnerStyles,
+                isActive,
+                expanded,
+                hasChild,
+                refItem,
+                provider
             );
-    }
+    },
 });