+++ src/components/Switch/types/index.d.ts
@@ -22,6 +22,7 @@ export declare type TBsSwitch = ComponentObjectPropsOptions<TSwitchOptionProps>;
 declare type AllowedSwitchProps = AllowedComponentProps & ComponentCustomProps &
     VNodeProps & EventUpdateModelValueProps<string | number | boolean> & {
     onChecked?: (checked: boolean) => void;
+    '@checked'?: (checked: boolean) => void;
 }
 
 export declare const BsSwitch: {